+++ lib/ui/components/reports/lists/types.ts
@@ -10,6 +10,8 @@ export interface ReportListItem {
 	totalTime: number;
 	hex?: string;
 	count?: number;
+	/** Defines the ordering in the group. (e.g. derive from time) */
+	order?: number;
 }
 
 export interface ProjectSummaryItem {